+++ GP/PBoard.cpp
@@ -102,7 +102,7 @@ CPlayBoard::CPlayBoard()
     m_crTextColor = RGB(0, 0, 0);
     m_crTextBoxColor = RGB(255, 255, 255);
     m_fontID = CGameBox::GetFontManager()->AddFont(
-        TenthPointsToScreenPixels(100), taBold, FF_SWISS, "Arial");
+        TenthPointsToScreenPixels(100), taBold, uint8_t(FF_SWISS), "Arial");
 
     m_bLockedDrawnBeneath = TRUE;
     m_bPVisible = TRUE;
